+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## v0.9.0 (unreleased)
 
+ * Support for DBN version 3:
+   * Handle DBN v3 headers
+   * Add `InstrumentDefMsgV3` and `StatMsgV3`
  * tui: Upgrade BubbleTea v2
  
 ## v0.8.10 (2026-03-22)